requirements.txt
requests
python-dotenv
fake-useragent
pyyaml
